+++ tests/chunking/test_line_based_document_chunker.py
@@ -316,7 +316,7 @@ class TestCompatibilityWithDocumentChunker:
 
     def test_has_same_interface_as_document_chunker(self):
         """Test that LineBasedDocumentChunker has the same public interface."""
-        from ingestion_pipeline.chunking.textract_document_chunker import TextractLayoutDocumentChunker
+        from ingestion_pipeline.chunking.textract_layout_chunker import TextractLayoutDocumentChunker
 
         line_chunker = LineBasedDocumentChunker()
         layout_chunker = TextractLayoutDocumentChunker(strategy_handlers={})
